refactor(main-dart): Ensures main.dart only handles entry point and MyApp setup.
